Relocate secret tags to a new file
